+++ app/src/main/java/ca/bcit/project/newwestnewlife/SplashActivity.java
@@ -1,0 +1,42 @@
+/*
+ * Created by Taran Rai Chris Kwon on 11/10/17 1:10 PM
+ * Copyright (c) 2017. All rights reserved
+ *
+ * Last modified 11/10/17 1:10 PM
+ */
+
+package ca.bcit.project.newwestnewlife;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.Window;
+
+public class SplashActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        this.requestWindowFeature(Window.FEATURE_NO_TITLE); //To get rid of the App name from Title bar
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_splash);
+
+        Thread welcomeThread = new Thread() {
+
+            @Override
+            public void run() {
+                try {
+                    super.run();
+                    sleep(3000);  //Delay of 3 seconds
+                } catch (Exception e) {
+
+                } finally {
+
+                    Intent i = new Intent(SplashActivity.this, HomepageActivity.class);
+                    startActivity(i);
+                    finish();
+                }
+            }
+        };
+        welcomeThread.start();
+    }
+}